requirements.txt
Django==1.2.1
GnuPGInterface==0.3.2
django-templatetag-sugar==0.1
flup==1.0.2
gearman==2.0.2
python-ldap==2.3.10
raven==2.0.0
simplejson==2.0.9
wsgiref==0.1.2
